Refactor: extract authentication.go
